Moved the RunBashCmdAsync function to a different file
