+++ cmd/utils/bash_commands.go
@@ -1,0 +1,23 @@
+package utils
+
+import (
+	"bytes"
+	"errors"
+	"os/exec"
+)
+
+func RunBashCmdAsync(cmd *exec.Cmd, printOutput func(), parseError func(errMsg string) string) {
+	var stderr bytes.Buffer
+	cmd.Stderr = &stderr
+	err := cmd.Start()
+	if err != nil {
+		errMsg := parseError(stderr.String())
+		PrettifyErrorIfExists(errors.New(errMsg))
+	}
+	printOutput()
+	err = cmd.Wait()
+	if err != nil {
+		errMsg := parseError(stderr.String())
+		PrettifyErrorIfExists(errors.New(errMsg))
+	}
+}
