+++ cmd/utils/error_handling.go
@@ -1,12 +1,8 @@
 package utils
 
 import (
-	"os"
-	"os/exec"
-
 	"github.com/fatih/color"
-	"errors"
-	"bytes"
+	"os"
 )
 
 func PrettifyErrorIfExists(err error) {
@@ -18,21 +14,5 @@ func PrettifyErrorIfExists(err error) {
 		}()
 		color.New(color.FgRed, color.Bold).Printf("💈 %s\n", err.Error())
 		panic(err)
-	}
-}
-
-func RunBashCmdAsync(cmd *exec.Cmd, printOutput func(), parseError func(errMsg string) string) {
-	var stderr bytes.Buffer
-	cmd.Stderr = &stderr
-	err := cmd.Start()
-	if err != nil {
-		errMsg := parseError(stderr.String())
-		PrettifyErrorIfExists(errors.New(errMsg))
-	}
-	printOutput()
-	err = cmd.Wait()
-	if err != nil {
-		errMsg := parseError(stderr.String())
-		PrettifyErrorIfExists(errors.New(errMsg))
 	}
 }
